Extracted method construction to MethodManagedFunctionBuilder
This allows re-use of reflection on languages that compile to Java to
enable similar rules for creating ManagedFunction instances from them.
